Extract exit-point detection into separate ExitPointDetection module
Move IsNoreturnCallDest, InstructionIsExitPoint template, and the three
IL-specific exit-point functions out of ReturnHighlightRenderLayer.cpp
into their own translation unit. This separates analysis logic from
render layer concerns and makes exit-point detection independently
testable.
